+++ src/components/atoms/icons/iconSvg.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { IIconSvg } from "@interfaces";
+import { IIconSvg } from "@interfaces/components";
 import { cn } from "@utilities";
 
 export const IconSvg = ({ className, alt = "icon", src: Svg, isVisible = true, disabled }: IIconSvg) => {
