+++ src/components/atoms/tabs/tabsContext.ts
@@ -1,7 +1,7 @@
 import React from "react";
-import { ITabsContext } from "@interfaces";
+import { ITabsContext } from "@interfaces/components";
 
 export const TabsContext = React.createContext<ITabsContext>({
 	activeTab: 0,
-	setActiveTab: () => {},
+	setActiveTab: () => null,
 });
