+++ src/enums/index.ts
@@ -1,4 +1,3 @@
 export { SessionStateType } from "@enums/sessionStates.enum";
 export { LoggerLevel } from "@enums/output.enum";
 export { SortOrder } from "@enums/sortOrder.enum";
-export { EButtonVariant } from "@enums/components";
