--- src/interfaces/index.ts
@@ -1,21 +1,0 @@
-export type {
-	IButton,
-	IDropdownButton,
-	IIconButton,
-	IBadge,
-	IDropdownMenu,
-	IFrame,
-	ILink,
-	ISubmenuInfo,
-	ISubmenu,
-	IMenu,
-	IConnectionsContent,
-	ITopbar,
-	ITab,
-	ITabList,
-	ITabs,
-	ITabsContext,
-	ITable,
-	IIcon,
-	IIconSvg,
-} from "@interfaces/components";
