+++ src/types/index.ts
@@ -1,2 +1,1 @@
 export type { ServiceResponse } from "@type/services.types";
-export type { TButtonVariant } from "@type/components";
